+++ NEWS.rst
@@ -4,6 +4,8 @@ News
 
 Version 0.8.2 - 2017-05-01
 
+  * NEW: Insert.delete_attrib(tag, ignore=False) - delete ATTRIB entities from the INSERT entity
+  * NEW: Insert.delete_all_attrib() - delete all ATTRIB entities from the INSERT entity
   * BUGFIX: setting attribs_follow=1 at INSERT entity before adding an attribute entity works
 
 Version 0.8.1 - 2017-04-06